apps list: better accessibility on switch toggles
now focusing the input will make screen readers correctly read the name
of the app and its description

replaced the Switch component by a new AppItem, which contains the whole
switch element + label + description of the app. This makes it easier to
read the code, instead of having half the code in the Apps component.