notice is still visible after clicking on "learn more" for better a11y:
if getting out of the modal without validating, it now allows keyboard
focus to go back on the "learn more" button.

If the button is invisible or not rendered in the DOM at all when
exiting the modal (like before), the script can't manage to focus back
the button when it shows up again.