notice is still visible after clicking on "learn more" for better a11y:
if getting out of the modal without validating, it allows keyboard focus
to go back on the "learn more" button. If the button is invisible/not
rendered in the DOM at all when exiting the modal, the script can't
manage to focus back the button when it shows up again.