Encode/decode cookie when saving/loading consents
Rationale for this change is this section from RFC2109
(https://tools.ietf.org/html/rfc2109):

```
Note: For backward compatibility, the separator in the Cookie header
is semi-colon (;) everywhere.  A server should also accept comma (,)
as the separator between cookie-values for future compatibility.
```

Several server-side implementations of Cookie Jars (e.g. rack, the
modular Ruby webserver interface, that is used in a.o. ruby-on-rails)
therefore split the `Cookie` header on `;` and `,`. The unescaped JSON
of the consents contains `,` which causes issues server-side.

This change is able to read cookies written by an older version as
`decodeURIComponent` will only replace escaped sequences.